organized files for django
