+++ .gitignore
@@ -1,1 +1,2 @@
 capstone-env
+.cache
