+++ FastAPI/creds.env
@@ -1,0 +1,2 @@
+client_id=17c2183917474c9ca53085e93a30bb83
+client_secret=fe0691cee0314448b731ffe483481bf5
