+++ FastAPI/requirements.txt
@@ -1,0 +1,42 @@
+anyio==3.6.1
+async-timeout==4.0.2
+certifi==2022.6.15
+charset-normalizer==2.0.12
+click==8.1.3
+colorama==0.4.5
+cycler==0.11.0
+Deprecated==1.2.13
+fastapi==0.78.0
+fonttools==4.33.3
+h11==0.13.0
+idna==3.3
+joblib==1.1.0
+kiwisolver==1.4.3
+matplotlib==3.5.2
+numpy==1.23.0
+packaging==21.3
+pandas==1.4.3
+Pillow==9.1.1
+plotly==5.9.0
+pydantic==1.9.1
+pyparsing==3.0.9
+python-dateutil==2.8.2
+python-dotenv==0.20.0
+python-multipart==0.0.5
+pytz==2022.1
+redis==4.3.4
+requests==2.28.0
+scikit-learn==1.1.1
+scipy==1.8.1
+seaborn==0.11.2
+six==1.16.0
+sklearn==0.0
+sniffio==1.2.0
+spotipy==2.20.0
+starlette==0.19.1
+tenacity==8.0.1
+threadpoolctl==3.1.0
+typing_extensions==4.2.0
+urllib3==1.26.9
+uvicorn==0.18.2
+wrapt==1.14.1
